+++ Chrome/src/offscreen/offscreen.js
@@ -10,7 +10,6 @@
     if (!msg || msg.type !== "peek:offscreen:look") return false;
 
     P.pipeline.look(msg.url, {
-      watchlist: msg.watchlist || [],
       images: !!msg.images
     })
       .then(sendResponse)